miner: remove special handling of hnspool support
Previously HNSPool supported multiple stratum protocols for HNS in an
attempt to help build a better stratum protocol. Due to lack of adoption
of this different spec, we have all but dropped support for the spec.
While technically it should work, we can't promise perfect operation as
most focus is spent on the more standard spec used by all other pools.

This commit simply removes the special support for HNSPool as we fully
support the more generic stratum spec. There is still likely a future in
which we implement this new spec, but it's not likely enough to warrant
keeping this special handling in handyminer.